Added email verification, restoring password

+++ src/main/java/com/itstep/ckj13_mvc/controller/NoteController.java
@@ -1,25 +1,28 @@
 package com.itstep.ckj13_mvc.controller;
 
 import com.itstep.ckj13_mvc.entity.Note;
+import com.itstep.ckj13_mvc.entity.User;
+import com.itstep.ckj13_mvc.repo.UserRepository;
 import com.itstep.ckj13_mvc.service.NoteService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 
 @Controller
 @RequestMapping("/notes")
 public class NoteController {
     private NoteService noteService;
-
-    public NoteController(NoteService noteService) {
-        this.noteService = noteService;
-    }
+    private UserRepository userRepository;
 
     @Autowired
-
+    public NoteController(NoteService noteService, UserRepository userRepository) {
+        this.noteService = noteService;
+        this.userRepository = userRepository;
+    }
 
     @GetMapping("/create")
     public String create() {
@@ -27,48 +30,84 @@ public class NoteController {
     }
 
     @PostMapping("/add")
-    public String add(@ModelAttribute(name = "note") Note note) {
+    public String add(@ModelAttribute(name = "note") Note note, Principal principal) {
+        User user = userRepository.findByUsername(principal.getName());
+        user.addNote(note);
         noteService.save(note);
+        userRepository.save(user);
         return "redirect:/notes";
     }
 
     @GetMapping
-    public String all(Model model) {
-        model.addAttribute("notes", noteService.findAll());
+    public String all(Model model, Principal principal) {
+        User user = userRepository.findByUsername(principal.getName());
+        model.addAttribute("notes", user.getNotes());
         return "notes";
     }
 
     @GetMapping("/{id}")
-    public String info(@PathVariable(name = "id") int id, Model model) {
-        // ищем объект по ид
+    public String info(@PathVariable(name = "id") int id, Model model, Principal principal) {
+        User user = userRepository.findByUsername(principal.getName());
+        List<Note> notes = user.getNotes();
         Note note = noteService.findById(id);
-        model.addAttribute("note", note);
-        return "info";
+        if (notes.contains(note)) {
+            model.addAttribute("note", note);
+            return "info";
+        } else {
+            return "redirect:/notes";
+        }
     }
 
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable(name = "id") int id) {
-        noteService.deleteById(id);
-        return "redirect:/notes";
+    public String delete(@PathVariable(name = "id") int id, Principal principal) {
+        User user = userRepository.findByUsername(principal.getName());
+        List<Note> notes = user.getNotes();
+        Note note = noteService.findById(id);
+        if (notes.contains(note)) {
+            noteService.deleteById(id);
+            return "redirect:/notes";
+        } else {
+            return "redirect:/notes";
+        }
     }
 
     @GetMapping("/edit/{id}")
-    public String edit(@PathVariable(name = "id") int id, Model model) {
-        model.addAttribute("note", noteService.findById(id));
-        return "edit-note";
+    public String edit(@PathVariable(name = "id") int id, Model model, Principal principal) {
+        User user = userRepository.findByUsername(principal.getName());
+        List<Note> notes = user.getNotes();
+        Note note = noteService.findById(id);
+        if (notes.contains(note)) {
+            model.addAttribute("note", note);
+            return "edit-note";
+        } else {
+            return "redirect:/notes";
+        }
     }
+
     @PostMapping("/edit")
-    public String updateUser(Note note){
+    public String updateUser(Note note, Principal principal) {
+        User user = userRepository.findByUsername(principal.getName());
+        user.addNote(note);
         noteService.save(note);
+        userRepository.save(user);
         return "redirect:/notes";
     }
+
     @GetMapping("/search")
-    public String search(@RequestParam(name = "word") String word, Model model) {
-        List<Note> notes = noteService.search("%"+word+"%");
+    public String search(@RequestParam(name = "word") String word, Model model, Principal principal) {
+        User user = userRepository.findByUsername(principal.getName());
+        List<Note> userNotes = user.getNotes();
+        List<Note> notes = noteService.search("%" + word + "%");
+        for (int i = 0; i < notes.size(); i++) {
+            Note note = notes.get(i);
+            if (!userNotes.contains(note)) {
+                notes.remove(note);
+            }
+        }
         model.addAttribute("notes", notes);
         return "notes";
     }
 
-    }
+}
